fix bug where scores were mostly 0; fix a few signatures

+++ net/nnue.hh
@@ -7,8 +7,8 @@ public:
     int delta_evaluate(Fenboard &b, move_t move, int previous_score);
 
 private:
-    void add_remove_piece(const Fenboard &b, int colored_piece_type, bool remove, int piece_pos, matrix<1, 512, int16_t> &layer);
-    void recalculate_dense1_layer(const Fenboard &b, matrix<1, 512, int16_t> &layer);
-    int calculate_score(const matrix<1, 512, int16_t> &input_layer, Color side_to_play) const;
-    matrix<1, 512, int16_t> dense_1_layer;
+    void add_remove_piece(const Fenboard &b, int colored_piece_type, bool remove, int piece_pos, matrix<1, 512, int8_t> &layer);
+    void recalculate_dense1_layer(const Fenboard &b, matrix<1, 512, int8_t> &layer);
+    int calculate_score(const matrix<1, 512, int8_t> &input_layer, Color side_to_play) const;
+    matrix<1, 512, int8_t> dense_1_layer;
 };
